+++ schemas/forecast_transactions.py
@@ -1,10 +1,11 @@
 from pydantic import BaseModel
 from datetime import date
+from decimal import Decimal
 
 class ForecastTransaction(BaseModel):
     date: date
     name: str
-    amount: float
+    amount: Decimal
     source_account_id: int | None = None
     destination_account_id: int | None = None
 
@@ -12,7 +13,7 @@ class ForecastPoint(BaseModel):
     account_id: int
     account_name: str
     date: date
-    balance: float
+    balance: Decimal
     is_external: bool
-    amount_in: float
-    amount_out: float
+    amount_in: Decimal
+    amount_out: Decimal
